requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asyncio-dgram==2.1.2
attrs==23.2.0
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mcstatus==11.1.1
orjson==3.10.0
outcome==1.3.0.post0
packaging==24.0
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
ujson==5.9.0
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
